Rockford Airport (+ reverted RenderAll)
Old airports from Rockford will be kept in the `_Rockford Old` for historical reasons (the `_` is so it gets passed by the renderer). I know this revision of RenderAll is really crappy, but the website needs it so that all charts are dumped in the one directory. Also the most recent revision broke the imagemagick section.